+++ lib/site_prism/page.rb
@@ -42,18 +42,6 @@ module SitePrism
       end
     end
 
-    # Where a Capybara HTML fragment has been directly injected into `#load` as a block return this loaded fragment
-    # Where a page has been directly navigated to through traditional means (i.e. Selenium), return an instance of the
-    # current Capybara session (With all applicable methods)
-    #
-    # @return [Capybara::Node::Simple || Capybara::Session]
-    def page
-      @_page ||= begin
-        SitePrism::Deprecator.deprecate('Calling #page on a SitePrism::Page instance')
-        to_capybara_node
-      end
-    end
-
     # This scopes our calls inside Page correctly to the `Capybara::Session`
     #
     # @return [Capybara::Node::Simple || Capybara::Session]