+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 - **Required Ruby Version is now 2.7+**
   ([luke-hill])
 
+- Internal change: All DSL code has been segregated and moved around. No method name changes
+  ([luke-hill])
+
 - DSL Validation has moved `DSLValidator` -> `DSL::Validator` and is now implemented by default
   - To opt out of this set the environment key `SITEPRISM_DSL_VALIDATION_ENABLED="false"`
     ([luke-hill])